Move JDS config to a separate file
